NgsGFX Metal: Fix memory leaks
objc_foundation::NSString should not have passed to msg_send! directly
because that means we are transferring the ownership of NSString to the
callee, which is not actually what the callee expects.

transmute_copy is used to pass the pointer to NSString without
transferring its ownership. As a result, NSStrings created by
NSString::from_str are now released correctly when they go out of scope.