Add explicit `()` return types for msg_send!
`msg_send!` with an unconstrained return type used to be deduced to have
the return type `()`. This is no longer the case after the stabilization
of the `!` (never) type (https://github.com/rust-lang/rust/issues/48950),
and it'll be deduced to be `!`.

This commit adds explicit return types to preserve the old behavior.