NgsPF Viewport: Perform update whenever a new frame was submitted
There is a chicken-and-egg issue here: before the creation of the first
window, no update is done because there is no frame to update. However,
without an update we can't run the reconcilation to create the first
window.

This commit solves this issue by trigger an update whenever a new
producer frame is submitted.

This hasn't been a problem with the Metal backend where the event loop
is periodically woken up whether a window exists or not.