Ngs.Interop: Fix `NgscomMarshal.QueryInterface` to support a certain form of casting
`QueryInterface` is a method provided to allow casting a RCW to some
unrelated interface (i.e., an interface that is not a base interface of
the RCW). In the previous implementation of the method, the object was
internally converted to `INativeObject<T>` to see if it's a RCW. This
causes the method to fail when it's supposed not to because even if it's
a RCW, it doesn't support `INativeObject<T>` if `T` is an unrelated
interface.

The new implementation tries the managed cast first. If it doesn't work,
it tries `QueryInterface` *without* converting it to `INativeObject<T>`.

The same applies to `NgscomMarshal.QueryInterfaceOrNull`.