Ngs.UI: Implement `Ngs.UI.TableLayout`
The optimal layout is computed using the simplex algorithm. As it turned
out, there were multiple issues in this initial implementation. Firstly,
the big idea of formulating this problem as a linear program was wrong
as a linear objective function is only capable of choosing the extreme
solutions. Secondly, using a specialized numeric type (`RankedFloat`) to
implement the priority scheme did not work well.