ZanGFX Metal: Switch the Rust edition to 2018
This change caused the compilation to fail in a previous version of the
nightly Rust compiler for unknown reasons. The latest version that I
have does not have this problem:
`rustc 1.29.0-nightly (6a1c0637c 2018-07-23)`.

One problem (probably a bug in the compiler?) still remains after this:
An explicit `pub extern crate` is still needed to re-export a crate. If
`pub extern crate zangfx_metal_rs as metal` is removed, `ngspf_viewport`
no longer compiles.