ZanGFX Vulkan: Implement the new heap-resource binding model
A vulkan device memory object is wrapped by a newly introduced type
`VulkanMemory`. Every buffer and image bound to it retains a reference
to it. The reference is stored in `HeapBinding`, which is protected
against simultaneous accesses indirectly by a heap's state mutex, by
using `TokenCell`.

`HeapBinding` also contains a `SysTlsfRegion`, which is used by
`Heap::make_aliasable` to invalidate the resource's memory allocation.