ZanGFX Utils: Generalize `streamer::Stage*`'s source/destination types
First, they are no longer re-imports of `uploader::Stage*` and now have
unique definitions.

Their fields of type `&[u8]`, `&BufferRef` and `&ImageRef` were replaced
with generic type parameters which are constrained to `Borrow` providing
references of intended types. This means that `Stage*` can be `'static`,
which is essential for them to be useful as input of `Streamer`.

A drawback of this is that their constructor functions no longer can
accept arbitrary POD types as input data.