ZanGFX: Reduce the memory overhead of handles
This commit introduces a trait `StableVtable`, which provides a
canonicalized version of `TraitObject::vtable` that is suitable for
representing a type identity, eliminating the need for storing `TypeId`
in handle types and thus reducing the size of each of them by 8 bytes.

In addition, this commit makes `SmallBox::vtable` non-null, which allows
further space optimization when a handle is stored in `Option` or
a similar enumerate type.

This approach has the following shortcomings:

- The runtime overhead of down-casting is slightly increased owing to an
  additional indirection required to access the canonicalized `vtable`.
  Still, it's supposedly faster than `std::any::Any`.

- Storing a generic type in a handle is even harder due to the lack of
  generic static items (<https://github.com/rust-lang/rfcs/pull/1520>).
  This approach requires a static variable where the canonical `vtable`
  for a specific type is stored. A potential work-around would be
  providing a slow path for types that can't generate a stable `vtable`
  pointers.