ZanGFX: Get rid of obscure macro dependencies
After this commit, `zangfx_base::zangfx_impl_object` no longer requires
the users to import `zangfx_base::{interfaces, vtable_for}` to the
context on which `zangfx_impl_object!` is invoked.

This commit also switches the version of `query_interface` to the one
that supports Rust 2018 style macro imports in order to accomplish the
goal.

Fixes #52.