NgsGameGFX: Add non-panic error handling to the CPU task graph manager
This commit adds an ability to return `Result` to `taskman::Task`. It's
intended to be used to handle ZanGFX errors (`zangfx::base::Error`).

It works pretty much in the same way as panic handling, except that
an error value and a panic status are managed in separate cells so if
both of them occur the latter is always prioritized.