ZanGFX Vulkan: Fix stage flags mismatch on `vkCmdWaitEvents`
This commit fixes an invalid usage of the Vulkan API that happens when
a command pass waits on more than one fences and some of them are
updated with empty stage flags while the others are not.

Fence updates are translated to `vkCmdSetEvent`. During this process,
empty stage flags are translated to `PIPELINE_STAGE_TOP_OF_PIPE_BIT`.
Fence waits are coalesced into a single `vkCmdWaitEvents` call, during
which stage flags used to update fences are OR-ed together. Previously,
this OR operation was performed *before* converting the stage flags to
the Vulkan equivalence, causing `PIPELINE_STAGE_TOP_OF_PIPE_BIT` not to
be included in the final output.