Update `ash` to 0.26.2
This update involves a considerable amount of changes because of how
`ash` changed in the way the Vulkan type definitions are converted
to Rust.

`ash` 0.25.0 removed its API versioning system and
`zangfx_vulkan::AshDevice` now therefore serves little purpose.

In feature-wise, the notable change is that the swapchain APIs now can
return `VK_ERROR_SUBOPTIMAL_KHR` as a successful value, which means that
`acquire_next_image_khr` can return an image index while returning a
flag indicating if a swapchain is in a suboptimal state.
The current code might or might not be handling the suboptimal status in
the best way conceivable.