ZanGFX: Replace `ngsenumflags` with `bitflags` in `zangfx_base`
This commit marks the start of the transition to `bitflags`.

`bitflags` is maintained more actively than `enumflags` and appears to
be fairly a common choice (even `ash`, maintained by the same author as
`enumflags`, uses `enumflags`). Although `bitflags` wraps a type
definition with a macro, Racer seems to recognize this pattern so
the code completion of RLS/Racer will keep working.

We use the crate `flags-macro` as a universal replacement of
`ngsenumflags::flags` (which lets us write flags in a compact way like
`flags![SomeFlags::{A | B | ... | Z}]`).