ZanGFX Vulkan: Fix validation error with MoltenVK
This commit fixes the following validation error:

    pImageMemBarriers[0].srcAccessMask (0x1800) is not supported by
    srcStageMask (0x1). The Vulkan spec states: Each element of
    pMemoryBarriers, pBufferMemoryBarriers and pImageMemoryBarriers must
    not have any access flag included in its srcAccessMask member if
    that bit is not supported by any of the pipeline stages in
    srcStageMask, as specified in the table of supported access types.
    (https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/
    vkspec.html#VUID-vkCmdPipelineBarrier-pMemoryBarriers-01184)

When MoltenVK is detected, the backend replaces calls to
`cmd_wait_events` with calls to `cmd_pipeline_barrier` which is
implemented in a separate code path.

The code path that uses `cmd_pipeline_barrier` is designed for the cases
where image layout transitions are required before a command pass but no
fence waits are defined for the pass. In such cases, the source pipeline
stage parameter provided to `cmd_pipeline_barrier` can be assumed to be
`TOP_OF_PIPE`.

However, this assumption does not apply to other cases, which are the
focus of this issue. This commit deals with the issue by deriving the
source pipeline stage flags appropriately to match the source access
masks of image memory barriers.