NgsPF Viewport: Add `PortFrame`
This commit introduces a trait `PortFrame`, which is used to exchange
information and objects between the compositor and a port
implementation and is recreated for every frame. The methods
of `PortInstance` pertaining to a particular frame were moved to
`PortFrame`.

`PortFrame` can store a reference to the current frame's
`PresenterFrame`. This brings about a notable change that
`image_(usage|format|extents)` (in addition to `render`) can now use it
to read NgsPF properties from, e.g., the `Port` object it was created
from.