Stygian: Improve the throughput of `SpanPainter`
`f32` addition has latency of 3–5 cycles depending on processors, thus
the speed of the successive update of the current Z value is limited by
the data dependency.

This commit implements an alternate strategy where each Z value is
calculated based on the base Z value and the current pixel position.
`cvtsi2ss` has 4x worse throughput than `addss` does, but since it's
calculated independently on each pixel, this strategy actually provides
a better performance.

`u23_to_f32` is not used here because it seems to provide an inferior
performance for some reason.